Remove unused argument
It looks like it was intended that load.pl would be able to run what's
actually in calc_caches.php but it doesn't at the moment.

It would be good if we could get it to in the future because then you
wouldn't need to remember to run two commands to update the database.