Get PHP version before Rails
This helps a bit when moving to compare_post_static for tests that
change things in the DB as it allows us to copy old.html and use it
as a fixutre.

If we get the Rails page before the PHP page, we're comparing the
already changed page which is incorrect.