Always include the policy ID so helper doesn't change it
This was happening if 'dmp' wasn't set in the params but was set by
the user's active policy instead. The result was the form was posting
to a URL for one policy but the form structure was set up for another
policy so it would never work.

Fixes #335

Tests were erroneously modified in 0d2267574a6c989ca749191d3582d7e338cf14e0