Fix calculation of turnout so that it's not based on party turnout
If every member of a party didn't vote on a division, their members
were not counted in the total number of possible people that could
have voted. This is clearly not right.

This fixes that but does cause a little confusion as parties that
don't vote are currently not shown at all on the party summary
table. So, at first glance the numbers don't add up. Once we
show parties that don't vote at all it will make sense once more.