Always mark tied votes as failed
Normally votes in the lower house should never be tied and the parser
is being updated to reflect that: https://github.com/openaustralia/openaustralia-parser/issues/7

There is one vote in the lower house that is tied but required an
absolute majority to pass so it failed. This change allows that.