Link back to the division
Despite what I said in e933776 it's more useful to go to the division
when you click on this link.

If you're editing policies (the main reason for this screen) and you
add or remove one, when you click on the title link it would go "back"
to this same page which is definitely not useful.