Remove reference to settings file in readme and also add an example
`config/settings/development.yml` takes care of settings in
development so we don't need the reference in the readme.

Adding the example helps illustrate you need it in production (we
don't yet describe how to deploy it though).

I've also moved the "loading data" section to it's own place as it's
useful to be able to do in development as well.

Fixes #948