Expand current_on scope to include memberships finishing on date
In the Ukraine data it's common for someone to leave a faction one day
and start a new faction the next. In this case it's clear that this
scope should include memberships finishing on the specified date.

I'm not sure if this is backwards compatible. We should investigate
and test further before we decide to merge it.