Disable PHP regression tests
We're going to be making significant changes to the text of the site
so these tests aren't going to help us.

In the future we should properly translate the site and run the tests
against a predicable language. Or better yet, get rid of these tests
and replace them with good unit tests and integration tests.