Set the locale in a before action filter & allow it to be changed
The reason for this change is that Passenger didn't seem to be picking
up the locale setting from application.rb. I think this is because the
locale is reset on each request so I'm hoping this now works in
production.

This example comes from the Rails Guides. It also allows you to change
the locale on the fly with a parameter which seems sensible. If we
ever want to support a multilingual site then we'll need to extend
this method anyway so it seems like a good place to start.