Fix colours of policy positions, closes #49
Rather than change the classes and the way the css works, I've just
added a new local variable, "position", for each position, that passes
the English version of the position which is used to create the class
name for the css to pick up.

The means we don't have to change the css, which I think is preferable
because it’s more changes, and at multiple layers, to deal with.

Doing it this way is a bit of a hack, but saves re-engineering how this
works.