Fix vote table disclosure function so only intended party is shown
When you click/tap the 'show members' button in the division votes table
, to see how the members in a party voted on this division, instead of
just the members of that party being shown, all members are shown.

This bug is caused by the text for an MP’s party being blank when it is
added to the tables data attributes with whip.party.parameterize. This
means that all the disclosure buttons target cells with '.member-row-'
which is all cells.

There's no reason why we need to use party names for this function.
This commit replaces it with a number, the index of the party in the
loop, using each_with_index.

This fixes #61