Make site name smaller to fit on one line with the nav button on small screens
The name of the site is longer in Ukrainian than in English. This means
that it doesn't fit on the same line as the small screen navigation
toggle button on phone sized screens.

This makes the site title smaller on very small screens so that it the
title and button fit on one line at common phone screen sizes (around 320px)

Fixes #38