Remove the rebel list on divisions, this doesn't need to be called out
In the Rada rebel voters are *much* more common that in Australia’s
parliament. Often a vote will have 20+ rebellions. We added this list in
the Australian version because if someone rebels its a big deal and we
wanted to draw attention to it. It's more common here so the list just
ends up becoming noise that's in the way of reading the votes table.