Paginate divisions to keep page size down
There are over 3000 divisions a year in the Rada. This means the
division page is super slow. Currently in production in tales 20
seconds to load the page, which downloads 4MB of HTML. See #92.

This adds pagination so that only a small number of divisions are
rendered. I've kept it to 100 for this first pass as that means you can
find the last few days without having to work through the pagination.