Only use "current_on" to find the member when loading a vote
This is essentially going back to doing the dumb thing. When we were
looking for the party it was causing problems in the rest of the
application because it only expects a single member per person to
exist.

So this means we don't support faction changes in the middle of the
day (#22) but we'll be able to be clearer about what the problem is
(#111).