Increase readability of author projects footer section
On wide-screens it's currently far too low contrast to be read.

This commit makes it the same colours as it is on smaller screens.
This makes it readable without any interaction, which is much simpler
and easier for users.

Thanks to @brtrx for catching this.
Fixes https://github.com/openaustralia/publicwhip/issues/1086