Update method to maintain behaviour after rails upgrade
This method is meant to make the date written to the db the current
local time, but with the UTC timezone. After upgrading from Rails 4.1
to rails 4.2.7 it no longer worked properly. The time was being shifted
back by the local offset to UTC before being written.

This commit changes the behaviour to more explicitly rewrite the date in
the method so the correct behaviour is returned.