Allow policy names longer that 50 character (matches current behaviour)
This changes the validation to check for 100 characters, rather than 50,
which was set in 5c9ae382c15f92da4b5cb8775fb8d7f73313b58a

100 matches the database column constraint you can see in the
db/schema.rb. In production we have policies with names longer than 50
characters, so I've added a test to confirm that they are valid, as well
as the test for this new behaviour for providing useful feedback to
users, see https://github.com/openaustralia/publicwhip/issues/1101

Not I've also changed the tests to build the policies using factories,
so you don't have to worry about the other attributes required for a
valid policy in these tests.