Member header layouts made same for pages
This commit aims to fix the issue https://github.com/openaustralia/publicwhip/issues/1078
The space to the right of the members photo is wider on the member show page than the divisions index with member page.
The divisions page has a difference of 10px whereas the member show page has a difference of 15px. 
To maintain consistency and considering that a 10px difference is more appropriate, this commit changes the 15px difference on the member show page to 10px.