Recent Changes page header layout changed
This commit aims to fix the issue https://github.com/openaustralia/publicwhip/issues/989
The Recent Changes page https://theyvoteforyou.org.au/history contained two lines below the header, the upper one coming from it as it was a page-header, and the lower one coming from the list of recent changes.
The upper line looked unnecessary, so it was removed. 
Also, to make the design consistent with the other pages layout, the header layout was changed and made similar to other pages like https://theyvoteforyou.org.au/help/faq.
In this commit, the Recent Changes page does not contain the unnecessary line and also has a consistent (similar to other pages) header layout.