Clear padding for the paragraph
This commit aims to fix the issue https://github.com/openaustralia/publicwhip/issues/1095
The MP votes on policy page has a problem with the vertical alignment as mentioned in the issue. The introductory paragraph does not align vertically with the other part of the page, it is shifted to the right because of the use of col-sm-7 class of Bootstrap, which introduces a padding of 15px.
To override this, a new id is added to this paragraph with the left padding made 0 in this commit.