Use present tense for former member rebellions
This commit aims to fix the issue openaustralia#726
Initially, on the Member Show page, 'Rebels' was used for current and former members as well. For former members, it should be 'Rebelled' instead.
After this commit, 'Rebels' would be used for current members and 'Rebelled' for former members.
Code written following DRY principles.