Show a tantalising message if there's been no edits in the last week
Before this change the recent edits page showed a blank page if nothing
had happened in the last week. It now tries to entice the viewer into
contributing by reminding them that they'll be shown on the history page
if they make a contribution. It also links off to pages to help them do
that.