Require webmock required for existing test
spec/lib/check_resources_exists_spec.rb doesn't work without it.
This snuck under the radar because it's filename was missing "_spec"
and it wasn't running as part of the test suit.

When I fixed that in the previous commit it started failing, because it
was missing webmock, which was removed when we added VCR in 80b8b68—probably
because we couldn't tell it was needed.