Replace call to fixtures with factory
Its quite confusing that this file makes a call to the fixtures,
independent of the call to the fixtures in the actual spec files that
use this.

This fix feels a bit wrong as well, I think the better option would be
to parse the user you wanted to the method from the spec file, so that
all test setup could be contained and understood from the spec file.