Replace fixtures with factories
This is the first step towards removing the fixture data which makes our
whole test sweet overly complex, because we have Factories and fixtures
and our database cleaning doesn't work as you'd expect it to.

Creating and deleting all these records for every example slows down the
tests dramatically. I think once the fixtures are removed we'll be able
to find a faster way to run these tests.