Don't require mysql for a development install
New developers on a mac may not have mysql running locally, and this
shouldn't strictly be needed for local dev. So rather than having this
requirement present, move it into the :production group that that it's
only required if Rails.env is set to production.

Tweak docs to split out the install of mysql from the rest of the dev
setup.